requirements.txt
fastapi==0.116.0
uvicorn[standard]==0.35.0
jinja2==3.1.6
sqlalchemy==2.0.41
psycopg[binary]==3.2.13
alembic==1.16.4
apscheduler==3.10.4
python-dotenv==1.1.1
pydantic-settings==2.10.1
python-multipart==0.0.20
slowapi==0.1.9
redis==6.2.0
httpx==0.28.1
google-generativeai==0.8.5
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
cryptography==45.0.5
